+++ lib/rapporteur/rspec.rb
@@ -7,8 +7,18 @@ shared_examples_for 'a successful status response' do
     expect(response.response_code).to(eq(200))
   end
 
-  it 'responds with a JSON content header' do
-    expect(response.content_type).to(eq(Mime[:json]))
+  context 'the response headers' do
+    it 'contains a Content-Type JSON header' do
+      expect(response.content_type).to(eq(Mime[:json]))
+    end
+
+    it 'does not contain an ETag header' do
+      expect(response.headers).not_to have_key('ETag')
+    end
+
+    it 'contains a Cache-Control header which disables client caching' do
+      expect(response.headers.fetch('Cache-Control')).to eq('no-cache')
+    end
   end
 
   context 'the response payload' do
@@ -25,13 +35,25 @@ shared_examples_for 'an erred status response' do
     expect(response.response_code).to(eq(500))
   end
 
-  it 'responds with a JSON content header' do
-    expect(response.content_type).to(eq(Mime[:json]))
+  context 'the response headers' do
+    it 'contains a Content-Type JSON header' do
+      expect(response.content_type).to(eq(Mime[:json]))
+    end
+
+    it 'does not contain an ETag header' do
+      expect(response.headers).not_to have_key('ETag')
+    end
+
+    it 'contains a Cache-Control header which disables client caching' do
+      expect(response.headers.fetch('Cache-Control')).to eq('no-cache')
+    end
   end
 
-  it 'contains errors' do
-    expect(parsed_body).to(have_key('errors'))
-    expect(parsed_body.fetch('errors')).not_to(be_empty)
+  context 'the response payload' do
+    it 'contains errors' do
+      expect(parsed_body).to(have_key('errors'))
+      expect(parsed_body.fetch('errors')).not_to(be_empty)
+    end
   end
 end